crawler/pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "beautifulsoup4",
    "python-dotenv",
    "langchain-google-genai",
    "pydantic",
    "browser-use",
    "asyncio",
    "aiohttp",
    "logging",
    "urllib3",
    "playwright",
    "datetime"
]
